SMGValue with tests
Add a class to represent SMG values. In Java library, SMG values were
represented as Integer instances. This is not that straightforward in
C++ so a small class was created to represent SMG values. Right now, the
class only wraps an integer serving as value ID. The class also soaked
the original SMGValueFactory.getNewValue() method, which is implented as
a class method of SMGValue class.